refactor: correct the code by fixing small bugs and refactoring styles
