+++ src/components/advantages/advantages.module.scss
@@ -4,8 +4,9 @@
   display: flex;
   flex-direction: column;
   align-items: center;
-  margin-top: 60px;
+  margin-top: 100px;
 }
+
 .textContent {
   @include section_header;
 }
