+++ src/components/banner/banner.module.scss
@@ -8,8 +8,7 @@
   border-radius: 10px;
   border: $border;
   padding-top: 30px;
-  background: url('../../assets/img/svg/banner_bg_circle.svg') no-repeat;
-  background-position: 100% 101%;
+  background: url('../../assets/img/svg/banner_bg_circle.svg') no-repeat 100% 101%;
 
   h2 {
     @include header(88px);
@@ -33,6 +32,7 @@
     h2 {
       max-width: 600px;
     }
+
     img {
       margin-top: 30px;
       width: 80%;
@@ -46,6 +46,7 @@
       font-size: 26px;
       max-width: 250px;
     }
+
     img {
       margin-top: 10px;
     }
