+++ src/components/button/button.module.scss
@@ -1,3 +1,5 @@
+@import '../../scss/allVariables.scss';
+
 .button {
   font-size: 17px;
   text-transform: lowercase;
@@ -6,21 +8,21 @@
   border: none;
   border-radius: 5px;
   color: white;
-  background: #3e66fb;
-  margin: 25px 0px;
+  background: $link;
+  margin: 25px 0;
 
   &:hover {
     cursor: pointer;
-    background: #2f4dbc;
+    background: $hover;
   }
 
   &:active {
-    background: #1f337e;
+    background: $active;
   }
 
   &:disabled {
     cursor: default;
-    background: #cccccc;
-    color: #9e9e9f;
+    background: $disabled;
+    color: $disabledText;
   }
 }
