+++ src/components/button/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import styles from './button.module.scss';
-import { ButtonProps } from '../../utils/helpers/interface';
+import { ButtonProps } from '../../interfaces/formInputs';
 
 const Button: React.FC<ButtonProps> = ({ label }) => {
   return <button className={styles.button}>{label}</button>;
