+++ src/components/faq/faq.module.scss
@@ -1,4 +1,4 @@
-@import '../../scss/allVariables.scss';
+  @import '../../scss/allVariables.scss';
 
 .container {
   display: flex;
