+++ src/components/faq/question/question.module.scss
@@ -2,15 +2,15 @@
 
 .container {
   display: flex;
-  border-radius: 5px;
   padding: 30px;
   column-gap: 10px;
   border: 1.5px solid #fff;
   align-items: flex-start;
-  cursor: pointer;
-  border-radius: 0px 0px 5px 0px;
+  border-radius: 0 0 5px 0;
   border-right: 1.5px solid #e5e5e5;
   border-bottom: 1.5px solid #e5e5e5;
+  transition: all 0.3s ease;
+
   h5 {
     @include text;
     font-size: 16px;
@@ -27,7 +27,6 @@
   &:hover {
     border-radius: 5px;
     border: 1.5px solid #3e66fb;
-    background: #ebebeb;
   }
 
   img {
@@ -38,6 +37,7 @@
 @media (max-width: 1240px) {
   .container {
     width: 70%;
+
     p {
       max-width: 100%;
     }
