+++ src/components/footer/footer.module.scss
@@ -34,6 +34,7 @@ footer {
 .gh_logo {
   width: 40px;
 }
+
 @media (max-width: 500px) {
   .container {
     flex-direction: column;
