+++ src/components/goToCatalog/goToCatalog.module.scss
@@ -3,7 +3,7 @@
 .container {
   display: flex;
   justify-content: center;
-  margin: 45px 0 120px 0;
+  margin: 45px 0 100px 0;
 }
 
 .catalogBtn {
