+++ src/components/header/header.module.scss
@@ -24,6 +24,7 @@ a {
   column-gap: 12px;
   align-items: center;
   color: #706c68;
+
   a {
     @include button_white;
   }
