+++ src/components/header/nav/navigation.module.scss
@@ -2,10 +2,11 @@
   display: flex;
   list-style: none;
   font-size: 14px;
-  margin: 28px 0 20px 0;
+  margin: 28px 0;
   column-gap: 46px;
   color: #323232;
 }
+
 .center {
   justify-content: center;
   display: flex;
@@ -15,6 +16,7 @@
   column-gap: 46px;
   color: #323232;
 }
+
 @media (max-width: 1220px) {
   .nav {
     justify-content: center;
