+++ src/components/inputs/dateInput/index.tsx
@@ -1,19 +1,21 @@
 import React from 'react';
-import { InputProps } from '../../../utils/helpers/interface';
+import { InputProps } from '../../../interfaces/formInputs';
 import styles from './dateInput.module.scss';
+import getInputClasses from '../../../utils/helpers/functions/getInputClasses';
+import getLabelClasses from '../../../utils/helpers/functions/getLabelClasses';
 
 const DateInput: React.FC<InputProps> = ({
-  label: label,
-  type: type,
-  id: id,
-  placeholder: placeholder,
-  hookData: hookData,
-  errorMessage: errorMessage,
-  isValid: isValid,
+  label,
+  type,
+  id,
+  placeholder,
+  hookData,
+  errorMessage,
+  isValid,
 }): JSX.Element => {
   return (
     <div className={styles.field}>
-      <label className={getLabelClasses(isValid)} htmlFor={id}>
+      <label className={getLabelClasses(isValid, styles)} htmlFor={id}>
         {label}
       </label>
       <input
@@ -21,31 +23,11 @@ const DateInput: React.FC<InputProps> = ({
         id={id}
         placeholder={placeholder}
         {...hookData}
-        className={getInputClasses(isValid)}
+        className={getInputClasses(isValid, styles)}
       />
       {errorMessage && <div className={styles.errors}>{errorMessage}</div>}
     </div>
   );
 };
-
-function getLabelClasses(isValid: boolean | undefined): string {
-  if (isValid === undefined) {
-    return styles['default-label'];
-  } else {
-    if (isValid) {
-      return styles['valid-label'];
-    } else {
-      return styles['error-label'];
-    }
-  }
-}
-
-function getInputClasses(isValid: boolean | undefined): string {
-  if (isValid === undefined || isValid) {
-    return styles['default-input'];
-  } else {
-    return styles['error-input'];
-  }
-}
 
 export default DateInput;
