+++ src/components/inputs/passwordInput/passwordInput.module.scss
@@ -19,6 +19,12 @@
   cursor: pointer;
 }
 
+@media (max-width: 820px) {
+  .field {
+    width: 110% !important;
+  }
+}
+
 @media (max-width: 750px) {
   .password-input {
     label {
