+++ src/components/inputs/selectInput/selectInput.module.scss
@@ -1,4 +1,5 @@
 @import '../defaultInput.module.scss';
+@import '../../../scss/allVariables.scss';
 
 .field {
   position: relative;
@@ -6,7 +7,11 @@
 }
 
 .default-input {
-  color: #cdcccc;
+  color: $defaultInput;
+}
+
+.error-input {
+  color: $defaultInput;
 }
 
 .valid-input {
@@ -14,7 +19,7 @@
   position: relative;
   border: 0;
   border-radius: 5px;
-  outline: 1px solid #cfcfcf;
+  outline: 1px solid $defaultOutline;
   padding: 14px 50px 14px 16px;
   width: 100%;
   background: white;
@@ -24,7 +29,7 @@
 
 select {
   appearance: none;
-  background: white !important;
+  background: white;
   background-image: url("data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' width='30px' height='20px' viewBox='0 0 24 24' fill='none'%3E%3Cpath d='M6.1018 8C5.02785 8 4.45387 9.2649 5.16108 10.0731L10.6829 16.3838C11.3801 17.1806 12.6197 17.1806 13.3169 16.3838L18.8388 10.0731C19.5459 9.2649 18.972 8 17.898 8H6.1018Z' fill='%233E66FB'/%3E%3C/svg%3E") !important;
   background-repeat: no-repeat !important;
   background-position: right 0.5rem center !important;
