+++ src/components/inputs/textInput/index.tsx
@@ -1,20 +1,22 @@
 import React from 'react';
-import { InputProps } from '../../../utils/helpers/interface';
+import { InputProps } from '../../../interfaces/formInputs';
 import styles from './textInput.module.scss';
+import getLabelClasses from '../../../utils/helpers/functions/getLabelClasses';
+import getInputClasses from '../../../utils/helpers/functions/getInputClasses';
 
 const TextInput: React.FC<InputProps> = ({
-  label: label,
-  type: type,
-  id: id,
-  placeholder: placeholder,
-  hookData: hookData,
-  errorMessage: errorMessage,
-  isValid: isValid,
-  className: className,
+  label,
+  type,
+  id,
+  placeholder,
+  hookData,
+  errorMessage,
+  isValid,
+  className,
 }): JSX.Element => {
   return (
     <div className={styles.field}>
-      <label className={getLabelClasses(isValid)} htmlFor={id}>
+      <label className={getLabelClasses(isValid, styles)} htmlFor={id}>
         {label}
       </label>
       <input
@@ -22,31 +24,11 @@ const TextInput: React.FC<InputProps> = ({
         id={id}
         placeholder={placeholder}
         {...hookData}
-        className={getInputClasses(isValid) + ' ' + className}
+        className={getInputClasses(isValid, styles) + ' ' + className}
       />
       {errorMessage && <div className={styles.errors}>{errorMessage}</div>}
     </div>
   );
 };
-
-function getLabelClasses(isValid: boolean | undefined): string {
-  if (isValid === undefined) {
-    return styles['default-label'];
-  } else {
-    if (isValid) {
-      return styles['valid-label'];
-    } else {
-      return styles['error-label'];
-    }
-  }
-}
-
-function getInputClasses(isValid: boolean | undefined): string {
-  if (isValid === undefined || isValid) {
-    return styles['default-input'];
-  } else {
-    return styles['error-input'];
-  }
-}
 
 export default TextInput;
