+++ src/components/layout/layout.module.scss
@@ -1,4 +1,5 @@
 @import '../../scss/allVariables.scss';
+
 .App {
   display: flex;
   flex-direction: column;
@@ -6,12 +7,14 @@
   padding: 0 20px;
   margin: 0 auto;
 }
+
 main {
   margin: 0 auto;
   max-width: 1216px;
   flex: 1 1 auto;
   width: 100%;
 }
+
 .gray_bg {
   background-color: $gray;
 }
