+++ src/index.scss
@@ -24,10 +24,16 @@ body {
   margin: 0;
   padding: 0;
 }
+
 $border: 1px solid #cfcfcf;
 
 input,
 select {
   font-family: 'Ubuntu', sans-serif;
   font-size: 15px;
+}
+
+a,
+button {
+  transition: all 0.3s ease;
 }
