+++ src/pages/404_notFound/404_notFound.module.scss
@@ -3,6 +3,7 @@
   flex-direction: column;
   align-items: center;
   padding: 60px 20px 0;
+
   img {
     width: 80%;
   }
