+++ src/pages/index.tsx
@@ -1,6 +1,6 @@
 import Main from './main';
 import About from './about';
 import Login from './login';
-import Registartion from './registration';
+import Registration from './registration';
 import PageNotFound from './404_notFound';
-export { Main, About, Login, Registartion, PageNotFound };
+export { Main, About, Login, Registration, PageNotFound };
