+++ src/pages/login/login.module.scss
@@ -1,5 +1,7 @@
+@import '../../scss/allVariables.scss';
+
 .login-form {
-  padding: 0px 30px 0px;
+  padding: 0 30px 0;
 
   button {
     margin-top: 0;
@@ -7,42 +9,16 @@
 }
 
 .container {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
+  @include form_container;
   max-width: 820px;
-  margin: 0 auto;
-  margin-bottom: 100px;
 }
 
 .page-title {
-  margin: 50px 0px 0px;
-  font-size: 46px;
-  text-transform: uppercase;
-  padding: 0px 30px;
-  text-align: center;
+  @include forms_page_title;
 }
 
 .registration-link {
-  display: flex;
-  align-items: center;
-  margin: 20px 0px 40px;
-  font-size: 17px;
-
-  a {
-    margin-left: 10px;
-    color: #2a76f7;
-    font-weight: 500;
-    text-decoration: underline;
-
-    &:hover {
-      color: #2f4dbc;
-    }
-      
-    &:active {
-      color: #1f337e;
-    }
-  }
+  @include form_link;
 }
 
 .form-container {
@@ -51,7 +27,7 @@
   justify-content: center;
   align-items: center;
   border-radius: 5px;
-  outline: 1px solid #cfcfcf;
+  outline: 1px solid $defaultOutline;
   width: 94%;
   height: 340px;
   overflow: hidden;
@@ -64,7 +40,7 @@
   position: absolute;
   bottom: 20px;
   padding: 5px;
-  background: red;
+  background: $error;
   border-radius: 3px;
 }
 
@@ -77,7 +53,7 @@
     border-radius: 50%;
     width: 592px;
     height: 592px;
-    background: #3e66fb;
+    background: $link;
     top: -110px;
     right: 56%;
     z-index: -1;
@@ -90,7 +66,7 @@
 
 @media (max-width: 750px) {
   .page-title {
-    margin: 35px 0px 0px;
+    margin: 35px 0 0;
     font-size: calc(10px + 4vmin);
   }
 
@@ -105,14 +81,14 @@
 
   .form-container {
     width: 90%;
-    padding: 40px 0px;
+    padding: 40px 0;
     height: 100%;
     overflow: visible;
     margin-top: 20px;
   }
 
   .login-form {
-    padding: 0px 40px 0px;
+    padding: 0 40px 0;
     width: 100%;
   }
 
@@ -125,7 +101,7 @@
 
   .form-error {
     font-size: 11px;
-    bottom:  -60px;
-    margin: 0px 0px;
+    bottom: -60px;
+    margin: 0 0;
   }
 }
