+++ src/router/requireAuth/requiereAuth.tsx
@@ -1,5 +1,6 @@
 import { Navigate, Outlet } from 'react-router-dom';
 import { userAuth } from '../../hooks/user-auth';
+import React from 'react';
 
 const RequireAuth: React.FC = () => {
   const auth = userAuth();
