+++ src/scss/mixins.scss
@@ -73,7 +73,29 @@
   color: $active;
 }
 
-// textcontent
+@mixin form_link {
+  display: flex;
+  align-items: center;
+  margin: 20px 0 40px;
+  font-size: 17px;
+
+  a {
+    margin-left: 10px;
+    color: $link;
+    font-weight: 500;
+    text-decoration: underline;
+
+    &:hover {
+      color: $hover;
+    }
+
+    &:active {
+      color: $active;
+    }
+  }
+}
+
+// text content
 
 @mixin section_header {
   h2 {
@@ -86,4 +108,33 @@
     max-width: 732px;
     text-align: center;
   }
+}
+
+@mixin forms_page_title {
+  margin: 30px 0 0;
+  font-size: 46px;
+  text-transform: uppercase;
+  padding: 0 30px;
+  text-align: center;
+}
+
+// other
+
+@mixin form_container {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  margin: 0 auto 100px;
+}
+
+@mixin label {
+  position: absolute;
+  top: -10px;
+  left: 10px;
+  font-size: 14px;
+  font-weight: 400;
+  line-height: 140%;
+  background: white;
+  padding: 0 5px;
+  z-index: 1;
 }
