+++ src/scss/variables.scss
@@ -1,5 +1,13 @@
 $gray: #343234;
 $disabled: #a1a1a1;
+$disabledText: #9e9e9f;
 $link: #3e66fb;
 $hover: #2f4dbc;
 $active: #1f337e;
+$defaultInput: #cdcccc;
+$defaultOutline: #cfcfcf;
+$defaultLabel: #706c68;
+$error: #fb3e3e;
+$warning: #f67f00;
+$success: #06c13a;
+
