+++ src/types/formInputs.ts
@@ -1,0 +1,21 @@
+export type Fields =
+  | 'email'
+  | 'password'
+  | 'firstName'
+  | 'lastName'
+  | 'dateOfBirth'
+  | 'shippingCountry'
+  | 'billingCountry'
+  | 'shippingCity'
+  | 'billingCity'
+  | 'shippingStreet'
+  | 'billingStreet'
+  | 'shippingHouseNumber'
+  | 'billingHouseNumber'
+  | 'shippingApartment'
+  | 'billingApartment'
+  | 'shippingPostcode'
+  | 'billingPostcode'
+  | 'sameAddress'
+  | 'defaultBilling'
+  | 'defaultShipping';
