+++ src/utils/helpers/functions/createInputs/billingAddress/createBillingHouseNumberInput.tsx
@@ -1,9 +1,9 @@
 import { UseFormReturn } from 'react-hook-form';
-import { FormFields } from '../../../interface';
+import { FormFields } from '../../../../../interfaces/formInputs';
 import { TextInput } from '../../../../../components/inputs';
 
 export default function createBillingHouseNumberInput(
-  form: UseFormReturn<FormFields, unknown, undefined>,
+  form: UseFormReturn<FormFields>,
   className: string,
 ): JSX.Element {
   const {
